+++ src/EmailAddress.vue
@@ -1,22 +1,17 @@
 <template>
     <span>{{ formattedAddress }}</span>
 </template>
-<script lang="ts">
-import { defineComponent } from 'vue'
 
-export default defineComponent({
-    name: 'EmailAddress',
-    props: {
-        address: {
-            type: Object,
-            required: true,
-        },
-    },
-    computed: {
-        formattedAddress() {
-            const a = this.address || {}
-            return `${a.name || ''} ${a.address ? `<${a.address}>` : ''}`
-        },
-    },
+<script lang="ts" setup>
+import type { Address } from 'postal-mime'
+import { computed } from 'vue'
+
+const props = defineProps<{
+    address: Address
+}>()
+
+const formattedAddress = computed(() => {
+    const a = props.address || {}
+    return `${a.name || ''} ${a.address ? `<${a.address}>` : ''}`
 })
 </script>