+++ src/Footer.vue
@@ -1,0 +1,23 @@
+<template>
+    <footer
+        class="bg-accent text-accent-content border-t-accent-content/30 flex flex-row justify-between gap-2 border-t p-2 text-xs md:px-6"
+    >
+        <div class="flex flex-col gap-2">
+            <span
+                >Current version maintained with <i class="fas fa-heart"></i> by
+                <a class="link" href="https://alexanderschroeder.net/"
+                    >Alexander&nbsp;Krivács&nbsp;Schrøder</a
+                ></span
+            >
+            <span>
+                Original made with <i class="fas fa-heart"></i> by
+                <a class="link" href="https://marcinchmiel.com">Marcin&nbsp;Chmiel</a>
+            </span>
+        </div>
+        <div>
+            <a class="link flex items-center" href="https://github.com/ilyvion/s3abird"
+                ><i class="fab fa-github"></i>&nbsp;Contribute</a
+            >
+        </div>
+    </footer>
+</template>